Update SQL queries, removing individual download tables
The code removes individual download tables and integrates them into a single comprehensive table using check constraints. This update improves the organization and accessibility of the data by merging it into a singular, unified structure. These changes primarily impact the SQL queries used for data management within the application.
