+++ .sqlx/query-47bd2d872c0a0c4e7d975f8f9e10472d83a9e3edc5557b7a798228e90a8742e2.json
@@ -1,0 +1,20 @@
+{
+  "db_name": "PostgreSQL",
+  "query": "SELECT MAX(updated_at) AS updated_at\nFROM download",
+  "describe": {
+    "columns": [
+      {
+        "ordinal": 0,
+        "name": "updated_at",
+        "type_info": "Timestamptz"
+      }
+    ],
+    "parameters": {
+      "Left": []
+    },
+    "nullable": [
+      null
+    ]
+  },
+  "hash": "47bd2d872c0a0c4e7d975f8f9e10472d83a9e3edc5557b7a798228e90a8742e2"
+}
