+++ .sqlx/query-4996e2f1e38ae696866b081e4f3d424aa0f90afd74a897d32d35ea07aba8eb81.json
@@ -1,6 +1,6 @@
 {
   "db_name": "PostgreSQL",
-  "query": "INSERT INTO episode_download (provider, title, episode, decimal, version, extra, created_at, updated_at)\nVALUES ($1, $2, $3, $4, $5, $6, $7, $8)\nON CONFLICT DO NOTHING\nRETURNING id",
+  "query": "INSERT INTO download (variant, provider, title, start_index, end_index, created_at, updated_at)\nVALUES ('batch', $1, $2, $3, $4, $5, $6)\nON CONFLICT DO NOTHING\nRETURNING id",
   "describe": {
     "columns": [
       {
@@ -15,8 +15,6 @@
         "Varchar",
         "Int4",
         "Int4",
-        "Int4",
-        "Text",
         "Timestamptz",
         "Timestamptz"
       ]
@@ -25,5 +23,5 @@
       false
     ]
   },
-  "hash": "eedb200d54cdfa888411b20e654b4f75a0740a9939320ad65185e8303f17bb28"
+  "hash": "4996e2f1e38ae696866b081e4f3d424aa0f90afd74a897d32d35ea07aba8eb81"
 }
