--- .sqlx/query-4d6b4ed5a0466ebfed54134f1a1d6b797c03af68ab2f1b4e18430ef1d1f1348e.json
@@ -1,15 +1,0 @@
-{
-  "db_name": "PostgreSQL",
-  "query": "UPDATE movie_download\nSET updated_at = $2\nWHERE id = $1",
-  "describe": {
-    "columns": [],
-    "parameters": {
-      "Left": [
-        "Uuid",
-        "Timestamptz"
-      ]
-    },
-    "nullable": []
-  },
-  "hash": "4d6b4ed5a0466ebfed54134f1a1d6b797c03af68ab2f1b4e18430ef1d1f1348e"
-}
