--- .sqlx/query-80097bd3a7d71331e48e804a08163ca57d2167086d82805249d72419c34a0324.json
@@ -1,15 +1,0 @@
-{
-  "db_name": "PostgreSQL",
-  "query": "UPDATE episode_download\nSET updated_at = $2\nWHERE id = $1",
-  "describe": {
-    "columns": [],
-    "parameters": {
-      "Left": [
-        "Uuid",
-        "Timestamptz"
-      ]
-    },
-    "nullable": []
-  },
-  "hash": "80097bd3a7d71331e48e804a08163ca57d2167086d82805249d72419c34a0324"
-}
