--- .sqlx/query-80c8c66c0f195dd5c87e26f5f3bf879b0d8abc8bed7bc396dd416bc23e8fb4ce.json
@@ -1,20 +1,0 @@
-{
-  "db_name": "PostgreSQL",
-  "query": "SELECT MAX(updated_at) AS updated_at\nFROM (SELECT MAX(updated_at) AS updated_at\n      FROM batch_download\n      UNION\n      SELECT MAX(updated_at) AS updated_at\n      FROM episode_download\n      UNION\n      SELECT MAX(updated_at) AS updated_at\n      FROM movie_download) AS variants",
-  "describe": {
-    "columns": [
-      {
-        "ordinal": 0,
-        "name": "updated_at",
-        "type_info": "Timestamptz"
-      }
-    ],
-    "parameters": {
-      "Left": []
-    },
-    "nullable": [
-      null
-    ]
-  },
-  "hash": "80c8c66c0f195dd5c87e26f5f3bf879b0d8abc8bed7bc396dd416bc23e8fb4ce"
-}
