+++ .sqlx/query-82e8429d1eb675f1e3a78619ad9c946a06aed3ec4aa8d08e415ab5aa939d7713.json
@@ -1,6 +1,6 @@
 {
   "db_name": "PostgreSQL",
-  "query": "SELECT ed.id, ed.updated_at, array_agg(edr.resolution) as resolutions\nFROM episode_download ed\n         inner JOIN episode_download_resolution edr on ed.id = edr.episode_download_id\nWHERE ed.provider = $1\n  AND ed.title = $2\n  AND ed.episode = $3\n  AND COALESCE(ed.decimal, -1) = COALESCE($4, -1)\n  AND COALESCE(ed.version, -1) = COALESCE($5, -1)\n  AND COALESCE(ed.extra, '') = COALESCE($6, '')\nGROUP BY id",
+  "query": "SELECT d.id, d.updated_at, array_agg(dr.resolution) as resolutions\nFROM download d\n         inner JOIN download_resolution dr on d.id = dr.download_id\nWHERE d.variant = 'episode'\n  AND d.provider = $1\n  AND d.title = $2\n  AND d.episode = $3\n  AND COALESCE(d.decimal, -1) = COALESCE($4, -1)\n  AND COALESCE(d.version, -1) = COALESCE($5, -1)\n  AND COALESCE(d.extra, '') = COALESCE($6, '')\nGROUP BY id",
   "describe": {
     "columns": [
       {
@@ -35,5 +35,5 @@
       null
     ]
   },
-  "hash": "c1ab24d522271baacf2ce2cb0efb86777ff816d6a33861cfa80f1ac2745e101b"
+  "hash": "82e8429d1eb675f1e3a78619ad9c946a06aed3ec4aa8d08e415ab5aa939d7713"
 }
