--- .sqlx/query-a0c588feef80ac8e3523985d0203e8d34c349eaaf4f735dc5b750e5b6eb16882.json
@@ -1,20 +1,0 @@
-{
-  "db_name": "PostgreSQL",
-  "query": "INSERT INTO batch_download_resolution (batch_download_id, resolution, torrent, file_name, comments, magnet, created_at)\nVALUES ($1, $2, $3, $4, $5, $6, $7)\nON CONFLICT DO NOTHING",
-  "describe": {
-    "columns": [],
-    "parameters": {
-      "Left": [
-        "Uuid",
-        "Int2",
-        "Varchar",
-        "Text",
-        "Text",
-        "Text",
-        "Timestamptz"
-      ]
-    },
-    "nullable": []
-  },
-  "hash": "a0c588feef80ac8e3523985d0203e8d34c349eaaf4f735dc5b750e5b6eb16882"
-}
