--- .sqlx/query-a71c44453ffcdd2cdc83d35198da388626c9d17b33cb0aed95855141ede1abbf.json
@@ -1,20 +1,0 @@
-{
-  "db_name": "PostgreSQL",
-  "query": "INSERT INTO movie_download_resolution (movie_download_id, resolution, torrent, file_name, comments, magnet, created_at)\nVALUES ($1, $2, $3, $4, $5, $6, $7)\nON CONFLICT DO NOTHING",
-  "describe": {
-    "columns": [],
-    "parameters": {
-      "Left": [
-        "Uuid",
-        "Int2",
-        "Varchar",
-        "Text",
-        "Text",
-        "Text",
-        "Timestamptz"
-      ]
-    },
-    "nullable": []
-  },
-  "hash": "a71c44453ffcdd2cdc83d35198da388626c9d17b33cb0aed95855141ede1abbf"
-}
