--- .sqlx/query-c2d0e70939c36c450acaaf076f932b383c900037a49a6d3a184d3bb03fabee40.json
@@ -1,25 +1,0 @@
-{
-  "db_name": "PostgreSQL",
-  "query": "INSERT INTO movie_download (provider, title, created_at, updated_at)\nVALUES ($1, $2, $3, $4)\nON CONFLICT DO NOTHING\nRETURNING id",
-  "describe": {
-    "columns": [
-      {
-        "ordinal": 0,
-        "name": "id",
-        "type_info": "Uuid"
-      }
-    ],
-    "parameters": {
-      "Left": [
-        "Varchar",
-        "Varchar",
-        "Timestamptz",
-        "Timestamptz"
-      ]
-    },
-    "nullable": [
-      false
-    ]
-  },
-  "hash": "c2d0e70939c36c450acaaf076f932b383c900037a49a6d3a184d3bb03fabee40"
-}
