--- .sqlx/query-cc3d550f2941462cab3aed632789689ef8f24027a4e586cf1276f6b787a55c4a.json
@@ -1,20 +1,0 @@
-{
-  "db_name": "PostgreSQL",
-  "query": "INSERT INTO episode_download_resolution (episode_download_id, resolution, torrent, file_name, comments, magnet, created_at)\nVALUES ($1, $2, $3, $4, $5, $6, $7)\nON CONFLICT DO NOTHING",
-  "describe": {
-    "columns": [],
-    "parameters": {
-      "Left": [
-        "Uuid",
-        "Int2",
-        "Varchar",
-        "Text",
-        "Varchar",
-        "Text",
-        "Timestamptz"
-      ]
-    },
-    "nullable": []
-  },
-  "hash": "cc3d550f2941462cab3aed632789689ef8f24027a4e586cf1276f6b787a55c4a"
-}
