+++ .sqlx/query-d29c1903ac9ec9b74297d2552e53525f6c274a6cd0dd28ef9d5fad7c10bc59e1.json
@@ -1,6 +1,6 @@
 {
   "db_name": "PostgreSQL",
-  "query": "SELECT bd.id, bd.updated_at, ARRAY_AGG(bdr.resolution) AS resolutions\nFROM batch_download bd\n         JOIN batch_download_resolution bdr ON bd.id = bdr.batch_download_id\nWHERE provider = $1\n  AND bd.title = $2\n  AND bd.start_index = $3\n  AND bd.end_index = $4\nGROUP BY id",
+  "query": "SELECT d.id, d.updated_at, ARRAY_AGG(dr.resolution) AS resolutions\nFROM download d\n         JOIN download_resolution dr ON d.id = dr.download_id\nWHERE d.variant = 'batch'\n  AND d.provider = $1\n  AND d.title = $2\n  AND d.start_index = $3\n  AND d.end_index = $4\nGROUP BY id",
   "describe": {
     "columns": [
       {
@@ -33,5 +33,5 @@
       null
     ]
   },
-  "hash": "ecfe5e6023e1580e26dfd711e71bbafb3919909c0edaff9bb0c2cb209074a61d"
+  "hash": "d29c1903ac9ec9b74297d2552e53525f6c274a6cd0dd28ef9d5fad7c10bc59e1"
 }
