+++ .sqlx/query-d4e7b1bbb0ca01ba8394a7c648e7d2bb988d60cbc5f33e3118755f808cc908bf.json
@@ -1,0 +1,29 @@
+{
+  "db_name": "PostgreSQL",
+  "query": "INSERT INTO download (variant, provider, title, episode, decimal, version, extra, created_at, updated_at)\nVALUES ('episode', $1, $2, $3, $4, $5, $6, $7, $8)\nON CONFLICT DO NOTHING\nRETURNING id",
+  "describe": {
+    "columns": [
+      {
+        "ordinal": 0,
+        "name": "id",
+        "type_info": "Uuid"
+      }
+    ],
+    "parameters": {
+      "Left": [
+        "Varchar",
+        "Varchar",
+        "Int4",
+        "Int4",
+        "Int4",
+        "Text",
+        "Timestamptz",
+        "Timestamptz"
+      ]
+    },
+    "nullable": [
+      false
+    ]
+  },
+  "hash": "d4e7b1bbb0ca01ba8394a7c648e7d2bb988d60cbc5f33e3118755f808cc908bf"
+}
