+++ .sqlx/query-e01ad738680138550ce7503a68b9029afb2f6fb851d9eed1426503ff5814ba23.json
@@ -1,0 +1,20 @@
+{
+  "db_name": "PostgreSQL",
+  "query": "INSERT INTO download_resolution (download_id, resolution, torrent, file_name, comments, magnet, created_at)\nVALUES ($1, $2, $3, $4, $5, $6, $7)\nON CONFLICT DO NOTHING",
+  "describe": {
+    "columns": [],
+    "parameters": {
+      "Left": [
+        "Uuid",
+        "Int2",
+        "Varchar",
+        "Text",
+        "Varchar",
+        "Text",
+        "Timestamptz"
+      ]
+    },
+    "nullable": []
+  },
+  "hash": "e01ad738680138550ce7503a68b9029afb2f6fb851d9eed1426503ff5814ba23"
+}
