+++ queries/batch/query_batch_download_by_unique.sql
@@ -1,8 +1,9 @@
-SELECT bd.id, bd.updated_at, ARRAY_AGG(bdr.resolution) AS resolutions
-FROM batch_download bd
-         JOIN batch_download_resolution bdr ON bd.id = bdr.batch_download_id
-WHERE provider = $1
-  AND bd.title = $2
-  AND bd.start_index = $3
-  AND bd.end_index = $4
+SELECT d.id, d.updated_at, ARRAY_AGG(dr.resolution) AS resolutions
+FROM download d
+         JOIN download_resolution dr ON d.id = dr.download_id
+WHERE d.variant = 'batch'
+  AND d.provider = $1
+  AND d.title = $2
+  AND d.start_index = $3
+  AND d.end_index = $4
 GROUP BY id
