--- queries/episode/update_episode_download_updated_at.sql
@@ -1,3 +1,0 @@
-UPDATE episode_download
-SET updated_at = $2
-WHERE id = $1
