+++ queries/movie/insert_movie_download.sql
@@ -1,4 +1,4 @@
-INSERT INTO movie_download (provider, title, created_at, updated_at)
-VALUES ($1, $2, $3, $4)
+INSERT INTO download (variant, provider, title, created_at, updated_at)
+VALUES ('movie', $1, $2, $3, $4)
 ON CONFLICT DO NOTHING
 RETURNING id
