+++ queries/movie/query_movie_download_by_unique.sql
@@ -1,6 +1,7 @@
-SELECT ed.id, ed.updated_at, ARRAY_AGG(edr.resolution) AS resolutions
-FROM movie_download ed
-         JOIN movie_download_resolution edr ON ed.id = edr.movie_download_id
-WHERE provider = $1
-  AND ed.title = $2
-GROUP BY id
+SELECT d.id, d.updated_at, ARRAY_AGG(dr.resolution) AS resolutions
+FROM download d
+         JOIN download_resolution dr ON d.id = dr.download_id
+WHERE d.variant = 'movie'
+  AND d.provider = $1
+  AND d.title = $2
+GROUP BY d.id
