--- queries/movie/update_movie_download_updated_at.sql
@@ -1,3 +1,0 @@
-UPDATE movie_download
-SET updated_at = $2
-WHERE id = $1
