+++ queries/query_last_updated_at.sql
@@ -1,9 +1,2 @@
 SELECT MAX(updated_at) AS updated_at
-FROM (SELECT MAX(updated_at) AS updated_at
-      FROM batch_download
-      UNION
-      SELECT MAX(updated_at) AS updated_at
-      FROM episode_download
-      UNION
-      SELECT MAX(updated_at) AS updated_at
-      FROM movie_download) AS variants
+FROM download
