+++ queries/update_download_updated_at.sql
@@ -1,3 +1,3 @@
-UPDATE batch_download
+UPDATE download
 SET updated_at = $2
 WHERE id = $1
